flask-services/src/borrowingdbservice/requirements.txt
#
# pip requirements for REST service catalogdbservice
#
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.18
protobuf==3.10.0
requests==2.22.0
six==1.13.0
urllib3==1.25.7
Werkzeug==0.16.0
